refactored page select into own component
