+++ client/components/index.js
@@ -15,4 +15,4 @@ export { default as UnmatchedRoute } from './UnmatchedRoute'
 export { default as SignupSuccess } from './SignupSuccess'
 export { default as SignupConfirm } from './SignupConfirm'
 export { default as SearchProductView } from './SearchProductView'
-
+export { default as PageSelector } from './PageSelector'
